New `tester` variant; better messaging for variants
The `tester` variant features a `--max N` flag that terminates the
server after N queries have been received.  This will be used for memory
and load testing, but shouldn't make it into production deployments.  As
such, messaging was added to the help screen for the FUZZ and TESTER
variants, in case those get out into the wild somehow.