Support `nsX.<base>` IN A queries transparently
When using glue records, the client re-resolves the nameserver records
against the IPs in the glue records (vs. just blindly trusting them), so
if netip.cc is it's own NS, it needs to respond to `ns1`, `ns2`, etc.,
with its own IP address.

Currently, this only supports IPv4.  Maybe in the future we can do v6