Ignore Additional Records in queries
Some clients (i.e. Linux, apparently) send OPT RRs in the Additional
Records section of the DNS query for things like "I don't do DNSSEC"
which was throwing off the sensitive re-use of the query datagram as the
reply.

We are now more forceful with our handling of the datagram buffer,
truncating the message after the first Question Record, and re-filling
the buffer from there.