Support Arbitrary A / NS replies via -a / -n
The new -a and -n (--ns) options let operators specify a split response,
not just the bound host, for A records (that are not "special") and NS
replies.  This also introduces support for numbered ns server A records
that are stable, i.e. ns1.netip.cc and ns2.netip.cc always return
different IPs, but each is "pinned" to the same reply.