ACME / Let's Encrypt! Support
netip.cc now has the "API" necessary to integrate with automatic
certificate enrolment through the ACME (RFC 8555) protocol, like
Let's Encrypt!  This only works with dns-01 challenges, for fairly
obvious reasons.

It works by abusing the fact that we get to place arbitrary rules on how
we interpret names and RR types.  To tell netip.cc about a new token for
a sub-domain, issue a specially-crafted IN A query:

  dig IN A _acme-challenge.YOUR.TOKEN.HERE._.foo.1.2.3.4.netip.cc

This will cause the netip.cc name server to put "YOURTOKENHERE" away in
a cache somewhere, and answer subsequent IN TXT requests with it:

  dig IN TXT _acme-challenge.foo.1.2.3.4.netip.cc

This is enough to prove domain ownership to ACME servers, for dns-01.

Challenge tokens are kept in memory for as long as it is feasible to do
so.  We implement a FIFO cache so that when we exhaust our cache space
for tokens, we evict the oldest token first.  This should work well in
practice, but is not resistant to an attacker trying to interfere
directly with ACME renewals.  The cache is server-wide.

This should make it easy for us to support both root-level certificates
and leaf (i.e. subdomain.W.X.Y.Z.netip.cc) certificates.

THIS IS HORRIBLY BROKEN AS WRITTEN, FROM A SECURITY STANDPOINT, AND
ALLOWS ANYONE TO IMPERSONATE NETIP.CC FROM AN HTTP PERSPECTIVE.